requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.35.0
django-appconf==1.0.2
django-bower==5.2.0
django-compressor==2.2
django-inline-svg==0.1.1
django-rest-framework==0.1.0
djangorestframework==3.7.7
idna==2.6
oauthlib==2.0.7
psycopg2==2.7.4
python3-openid==3.1.0
pytz==2018.3
rcssmin==1.0.6
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
six==1.11.0
urllib3==1.22
